% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autolgm-package.R
\docType{package}
\name{autolgm-package}
\alias{autolgm}
\alias{autolgm-package}
\title{autolgm: Tool For Comparing The Fit Indices of Various Possible Latent Growth Models Automatically}
\description{
This package will provide the fit indices results for the no-change (intercept-only), linear, quadratic, and piece-wise latent growth model (if you request) automatically with simple code. In addition, the results of the models with various level-1 residual structures will be provided.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Junyeong1130/autolgm}
  \item Report bugs at \url{https://github.com/Junyeong1130/autolgm/issues}
}

}
\author{
\strong{Maintainer}: Junyeong Yang \email{junyeong1130@gmail.com} (\href{https://orcid.org/0000-0003-2478-2492}{ORCID})

}
\keyword{internal}
